componentizando e remover comments
